+++ backend/src/objects/infos.rs
@@ -10,5 +10,5 @@ pub struct Infos {
     pub players: Vec<Player>,
     pub apples: Vec<Vector2>,
     pub size: Vector2,
-    pub message_count: u64,
+    pub message_count: u32,
 }
